+++ mangadex_library/lib/models/common/content_rating.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum ContentRating {
   safe,
   suggestive,