+++ mangadex_library/lib/models/common/future_updates.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum FutureUpdates {
   enable,
   disable,