+++ mangadex_library/lib/models/common/language_codes.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum LanguageCodes {
   en, // English
   zh, // Simplified Chinese