+++ mangadex_library/lib/models/common/manga_status.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum MangaStatus {
   ongoing,
   completed,