+++ mangadex_library/lib/models/common/publication_demographic.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum PublicDemographic {
   shounen,
   shoujo,