+++ mangadex_library/lib/models/common/reading_status.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum ReadingStatus {
   reading,
   on_hold,