+++ mangadex_library/lib/models/common/tag_modes.dart
@@ -1,4 +1,3 @@
-///@nodoc
 enum TagsMode {
   Or,
   And,