+++ mangadex_library/lib/models/common/visibility.dart
@@ -1,0 +1,4 @@
+enum Visibility {
+  private,
+  public,
+}